+++ src/pages/_app.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Head from 'next/head';
 import { ThemeProvider } from '@material-ui/core/styles';
-// import CssBaseline from '@material-ui/core/CssBaseline';
+import CssBaseline from '@material-ui/core/CssBaseline';
 import theme from '../theme';
 import Layout from '../components/Layout';
 // import probesService from '../ProbesService';
@@ -33,7 +33,7 @@ export default function MyApp(props) {
       </Head>
       <ThemeProvider theme={theme}>
         <Layout title="My Cocktails">
-          {/* <CssBaseline /> */}
+          <CssBaseline />
           <Component {...pageProps} />
         </Layout>
       </ThemeProvider>
