+++ src/pages/home.js
@@ -1,0 +1,16 @@
+import { useRouter } from 'next/router';
+import { useEffect } from 'react';
+
+function Redirect({ to }) {
+  const router = useRouter();
+
+  useEffect(() => {
+    router.push(to);
+  }, [to]);
+
+  return null;
+}
+
+export default function HomePage() {
+  return <Redirect to="/" />;
+}
